Expose callback API for replacing low-level cryptographic primitives
This makes the callback API to replace low-level cryptographic
implementation public again after open-quantum-safe#1667.
