+++ src/common/sha3/sha3.h
@@ -18,7 +18,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include <oqs/common.h>
+#include <oqs/sha3_ops.h>
 
 #if defined(__cplusplus)
 extern "C" {
@@ -39,12 +39,6 @@ extern "C" {
  * \param inplen The number of message bytes to process
  */
 void OQS_SHA3_sha3_256(uint8_t *output, const uint8_t *input, size_t inplen);
-
-/** Data structure for the state of the incremental SHA3-256 API. */
-typedef struct {
-	/** Internal state. */
-	void *ctx;
-} OQS_SHA3_sha3_256_inc_ctx;
 
 /**
  * \brief Initialize the state for the incremental SHA3-256 API.
@@ -115,12 +109,6 @@ void OQS_SHA3_sha3_256_inc_ctx_clone(OQS_SHA3_sha3_256_inc_ctx *dest, const OQS_
  */
 void OQS_SHA3_sha3_384(uint8_t *output, const uint8_t *input, size_t inplen);
 
-/** Data structure for the state of the incremental SHA3-384 API. */
-typedef struct {
-	/** Internal state. */
-	void *ctx;
-} OQS_SHA3_sha3_384_inc_ctx;
-
 /**
  * \brief Initialize the state for the incremental SHA3-384 API.
  *
@@ -189,12 +177,6 @@ void OQS_SHA3_sha3_384_inc_ctx_clone(OQS_SHA3_sha3_384_inc_ctx *dest, const OQS_
  * \param inplen The number of message bytes to process
  */
 void OQS_SHA3_sha3_512(uint8_t *output, const uint8_t *input, size_t inplen);
-
-/** Data structure for the state of the incremental SHA3-512 API. */
-typedef struct {
-	/** Internal state. */
-	void *ctx;
-} OQS_SHA3_sha3_512_inc_ctx;
 
 /**
  * \brief Initialize the state for the incremental SHA3-512 API.
@@ -267,12 +249,6 @@ void OQS_SHA3_sha3_512_inc_ctx_clone(OQS_SHA3_sha3_512_inc_ctx *dest, const OQS_
  * \param inplen The number of seed bytes to process
  */
 void OQS_SHA3_shake128(uint8_t *output, size_t outlen, const uint8_t *input, size_t inplen);
-
-/** Data structure for the state of the incremental SHAKE-128 API. */
-typedef struct {
-	/** Internal state. */
-	void *ctx;
-} OQS_SHA3_shake128_inc_ctx;
 
 /**
  * \brief Initialize the state for the incremental SHAKE-128 API.
@@ -355,12 +331,6 @@ void OQS_SHA3_shake128_inc_ctx_reset(OQS_SHA3_shake128_inc_ctx *state);
  */
 void OQS_SHA3_shake256(uint8_t *output, size_t outlen, const uint8_t *input, size_t inplen);
 
-/** Data structure for the state of the incremental SHAKE-256 API. */
-typedef struct {
-	/** Internal state. */
-	void *ctx;
-} OQS_SHA3_shake256_inc_ctx;
-
 /**
  * \brief Initialize the state for the incremental SHAKE-256 API.
  *
@@ -422,206 +392,6 @@ void OQS_SHA3_shake256_inc_ctx_clone(OQS_SHA3_shake256_inc_ctx *dest, const OQS_
  * \param state The function state; must be initialized
  */
 void OQS_SHA3_shake256_inc_ctx_reset(OQS_SHA3_shake256_inc_ctx *state);
-
-/** Data structure implemented by cryptographic provider for SHA-3 operations.
- */
-struct OQS_SHA3_callbacks {
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256.
-	 */
-	void (*SHA3_sha3_256)(uint8_t *output, const uint8_t *input, size_t inplen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_init.
-	 */
-	void (*SHA3_sha3_256_inc_init)(OQS_SHA3_sha3_256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_absorb.
-	 */
-	void (*SHA3_sha3_256_inc_absorb)(OQS_SHA3_sha3_256_inc_ctx *state, const uint8_t *input, size_t inlen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_finalize.
-	 */
-	void (*SHA3_sha3_256_inc_finalize)(uint8_t *output, OQS_SHA3_sha3_256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_ctx_release.
-	 */
-	void (*SHA3_sha3_256_inc_ctx_release)(OQS_SHA3_sha3_256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_ctx_reset.
-	 */
-	void (*SHA3_sha3_256_inc_ctx_reset)(OQS_SHA3_sha3_256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_256_inc_ctx_clone.
-	 */
-	void (*SHA3_sha3_256_inc_ctx_clone)(OQS_SHA3_sha3_256_inc_ctx *dest, const OQS_SHA3_sha3_256_inc_ctx *src);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384.
-	 */
-	void (*SHA3_sha3_384)(uint8_t *output, const uint8_t *input, size_t inplen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_init.
-	 */
-	void (*SHA3_sha3_384_inc_init)(OQS_SHA3_sha3_384_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_absorb.
-	 */
-	void (*SHA3_sha3_384_inc_absorb)(OQS_SHA3_sha3_384_inc_ctx *state, const uint8_t *input, size_t inlen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_finalize.
-	 */
-	void (*SHA3_sha3_384_inc_finalize)(uint8_t *output, OQS_SHA3_sha3_384_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_ctx_release.
-	 */
-	void (*SHA3_sha3_384_inc_ctx_release)(OQS_SHA3_sha3_384_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_ctx_reset.
-	 */
-	void (*SHA3_sha3_384_inc_ctx_reset)(OQS_SHA3_sha3_384_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_384_inc_ctx_clone.
-	 */
-	void (*SHA3_sha3_384_inc_ctx_clone)(OQS_SHA3_sha3_384_inc_ctx *dest, const OQS_SHA3_sha3_384_inc_ctx *src);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512.
-	 */
-	void (*SHA3_sha3_512)(uint8_t *output, const uint8_t *input, size_t inplen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_init.
-	 */
-	void (*SHA3_sha3_512_inc_init)(OQS_SHA3_sha3_512_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_absorb.
-	 */
-	void (*SHA3_sha3_512_inc_absorb)(OQS_SHA3_sha3_512_inc_ctx *state, const uint8_t *input, size_t inlen);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_finalize.
-	 */
-	void (*SHA3_sha3_512_inc_finalize)(uint8_t *output, OQS_SHA3_sha3_512_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_ctx_release.
-	 */
-	void (*SHA3_sha3_512_inc_ctx_release)(OQS_SHA3_sha3_512_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_ctx_reset.
-	 */
-	void (*SHA3_sha3_512_inc_ctx_reset)(OQS_SHA3_sha3_512_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_sha3_512_inc_ctx_clone.
-	 */
-	void (*SHA3_sha3_512_inc_ctx_clone)(OQS_SHA3_sha3_512_inc_ctx *dest, const OQS_SHA3_sha3_512_inc_ctx *src);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128.
-	 */
-	void (*SHA3_shake128)(uint8_t *output, size_t outlen, const uint8_t *input, size_t inplen);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_init.
-	 */
-	void (*SHA3_shake128_inc_init)(OQS_SHA3_shake128_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_absorb.
-	 */
-	void (*SHA3_shake128_inc_absorb)(OQS_SHA3_shake128_inc_ctx *state, const uint8_t *input, size_t inlen);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_finalize.
-	 */
-	void (*SHA3_shake128_inc_finalize)(OQS_SHA3_shake128_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_squeeze.
-	 */
-	void (*SHA3_shake128_inc_squeeze)(uint8_t *output, size_t outlen, OQS_SHA3_shake128_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_ctx_release.
-	 */
-	void (*SHA3_shake128_inc_ctx_release)(OQS_SHA3_shake128_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_ctx_clone.
-	 */
-	void (*SHA3_shake128_inc_ctx_clone)(OQS_SHA3_shake128_inc_ctx *dest, const OQS_SHA3_shake128_inc_ctx *src);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake128_inc_ctx_reset.
-	 */
-	void (*SHA3_shake128_inc_ctx_reset)(OQS_SHA3_shake128_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256.
-	 */
-	void (*SHA3_shake256)(uint8_t *output, size_t outlen, const uint8_t *input, size_t inplen);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_init.
-	 */
-	void (*SHA3_shake256_inc_init)(OQS_SHA3_shake256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_absorb.
-	 */
-	void (*SHA3_shake256_inc_absorb)(OQS_SHA3_shake256_inc_ctx *state, const uint8_t *input, size_t inlen);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_finalize.
-	 */
-	void (*SHA3_shake256_inc_finalize)(OQS_SHA3_shake256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_squeeze.
-	 */
-	void (*SHA3_shake256_inc_squeeze)(uint8_t *output, size_t outlen, OQS_SHA3_shake256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_ctx_release.
-	 */
-	void (*SHA3_shake256_inc_ctx_release)(OQS_SHA3_shake256_inc_ctx *state);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_ctx_clone.
-	 */
-	void (*SHA3_shake256_inc_ctx_clone)(OQS_SHA3_shake256_inc_ctx *dest, const OQS_SHA3_shake256_inc_ctx *src);
-
-	/**
-	 * Implementation of function OQS_SHA3_shake256_inc_ctx_reset.
-	 */
-	void (*SHA3_shake256_inc_ctx_reset)(OQS_SHA3_shake256_inc_ctx *state);
-};
-
-/**
- * Set callback functions for SHA3 operations.
- *
- * This function may be called before OQS_init to switch the
- * cryptographic provider for SHA3 operations. If it is not called,
- * the default provider determined at build time will be used.
- *
- * @param new_callbacks Callback functions defined in OQS_SHA3_callbacks struct
- */
-OQS_API void OQS_SHA3_set_callbacks(struct OQS_SHA3_callbacks *new_callbacks);
 
 #if defined(__cplusplus)
 } // extern "C"
